ai_model/requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
kagglehub==0.2.7
pillow==10.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
